requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argcomplete==3.0.5
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blosc2==2.0.0
boto==2.49.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.0.4
ci-info==0.3.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.3.9
comm==0.1.2
configobj==5.0.8
configparser==5.3.0
contourpy==1.0.7
crcmod==1.7
cryptography==38.0.4
cupy-cuda117==10.6.0
cycler==0.11.0
Cython==0.29.33
debugpy==1.5.1
decorator==5.1.1
deepctr==0.9.3
defusedxml==0.7.1
entrypoints==0.4
etelemetry==0.3.0
executing==0.8.3
fastdtw==0.3.4
fasteners==0.18
fastjsonschema==2.17.1
fastrlock==0.8.1
filelock==3.9.0
fitz==0.0.1.dev2
flanker==0.9.11
flatbuffers==23.1.21
flit_core==3.6.0
fonttools==4.38.0
fqdn==1.5.1
frontend==0.0.3
frozenlist==1.3.3
fst-pso==1.8.1
future==0.18.3
fuxictr==2.0.1
FuzzyTM==2.0.5
gast==0.4.0
gcs-oauth2-boto-plugin==3.0
ge==0.0.0
gensim==4.3.0
google-apitools==0.5.32
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-reauth==0.1.1
graphembedding==0.1
grpcio==1.51.1
gsutil==5.23
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.8.0
hdbscan==0.8.29
httplib2==0.20.4
huggingface-hub==0.13.4
hyperopt==0.2.7
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==7.4.8
jupyter-console==6.6.3
jupyter_core==5.1.1
jupyter-events==0.6.3
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.11.0
kiwisolver==1.4.4
lab==7.3
learn2learn==0.1.6
libclang==15.0.6.1
lightgbm==3.3.5
llvmlite==0.39.1
lmdb==1.4.0
looseversion==1.0.3
lxml==4.9.2
marisa-trie==0.7.8
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
miniful==0.0.6
mistune==3.0.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
monotonic==1.6
msgpack==1.0.4
multidict==6.0.4
mwparserfromhell==0.6.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==2.8.8
nibabel==5.0.0
nipype==1.8.5
nltk==3.8.1
node2vec==0.4.6
notebook==6.5.4
notebook_shim==0.2.3
numba==0.56.4
numexpr==2.8.4
numpy>=1.19.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.0
pandas>=1.2.0
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
platformdirs==2.5.2
ply==3.11
prometheus-client==0.17.0
prompt-toolkit==3.0.36
protobuf==3.19.6
prov==2.0.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pyFUME==0.2.25
Pygments==2.11.2
pynndescent==0.5.8
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-version==0.0.2
pythonwhois==2.4.3
pytrec-eval==0.5
pytz==2022.7.1
pyu2f==0.1.5
pyxnat==1.5
PyYAML>=5.4.1
pyzmq==25.1.0
qpth==0.0.15
qtconsole==5.4.3
QtPy==2.3.1
rdflib==6.2.0
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
retry-decorator==1.1.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.7.2
sacremoses==0.0.53
SciencePlots==2.0.1
scikit-learn>=0.24.0
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
sentencepiece==0.1.98
setuptools==65.6.3
simpful==2.9.0
simplejson==3.18.3
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.2.0
starlette==0.24.0
tables==3.8.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-text==2.11.0
tensorrt==8.5.2.2
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tld==0.13
tldextract==3.4.0
tokenizers==0.13.3
top2vec==1.0.29
torch>=1.9.0
torch-tb-profiler==0.4.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm>=4.62.0
traitlets==5.7.1
traits==6.3.2
transformers==4.28.1
txt2tags==3.8
typing_extensions==4.4.0
umap-learn==0.5.3
uri-template==1.3.0
urllib3==1.26.14
uvicorn==0.20.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.6.0
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.7
wordcloud==1.8.2.2
wordnet==0.0.1b2
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
flwr>=1.4.0
